suppression controleur parentcontroller et intégration update dans controleur profilcontroller
